+++ sdk/js/src/utils/array.ts
@@ -9,9 +9,12 @@ import {
 import { canonicalAddress, humanAddress, isNativeDenom } from "../terra";
 import {
   ChainId,
+  ChainName,
   CHAIN_ID_ALGORAND,
+  CHAIN_ID_NEAR,
   CHAIN_ID_SOLANA,
   CHAIN_ID_TERRA,
+  coalesceChainId,
   isEVMChain,
 } from "./consts";
 
@@ -50,10 +53,12 @@ export const hexToUint8Array = (h: string): Uint8Array =>
  *
  * @throws if address is not the right length for the given chain
  */
+
 export const tryUint8ArrayToNative = (
   a: Uint8Array,
-  chainId: ChainId
+  chain: ChainId | ChainName
 ): string => {
+  const chainId = coalesceChainId(chain);
   if (isEVMChain(chainId)) {
     return hexZeroPad(hexValue(a), 20);
   } else if (chainId === CHAIN_ID_SOLANA) {
@@ -115,8 +120,10 @@ export const hexToNativeAssetString = (
  *
  * @throws if address is not the right length for the given chain
  */
-export const tryHexToNativeString = (h: string, c: ChainId): string =>
-  tryUint8ArrayToNative(hexToUint8Array(h), c);
+export const tryHexToNativeString = (
+  h: string,
+  c: ChainId | ChainName
+): string => tryUint8ArrayToNative(hexToUint8Array(h), c);
 
 /**
  *
@@ -127,7 +134,7 @@ export const tryHexToNativeString = (h: string, c: ChainId): string =>
  */
 export const hexToNativeString = (
   h: string | undefined,
-  c: ChainId
+  c: ChainId | ChainName
 ): string | undefined => {
   if (!h) {
     return undefined;
@@ -149,13 +156,14 @@ export const hexToNativeString = (
  */
 export const tryNativeToHexString = (
   address: string,
-  chain: ChainId
+  chain: ChainId | ChainName
 ): string => {
-  if (isEVMChain(chain)) {
+  const chainId = coalesceChainId(chain);
+  if (isEVMChain(chainId)) {
     return uint8ArrayToHex(zeroPad(arrayify(address), 32));
-  } else if (chain === CHAIN_ID_SOLANA) {
+  } else if (chainId === CHAIN_ID_SOLANA) {
     return uint8ArrayToHex(zeroPad(new PublicKey(address).toBytes(), 32));
-  } else if (chain === CHAIN_ID_TERRA) {
+  } else if (chainId === CHAIN_ID_TERRA) {
     if (isNativeDenom(address)) {
       return (
         "01" +
@@ -168,10 +176,13 @@ export const tryNativeToHexString = (
     }
   } else if (chain === CHAIN_ID_ALGORAND) {
     return nativeStringToHexAlgorand(address);
+  } else if (chainId === CHAIN_ID_NEAR) {
+    // TODO: handle algorand
+    throw Error("hexToNativeString: Near not supported yet.");
   } else {
     // If this case is reached
-    const _: never = chain;
-    throw Error("Don't know how to convert address from chain " + chain);
+    const _: never = chainId;
+    throw Error("Don't know how to convert address from chain " + chainId);
   }
 };
 
@@ -185,7 +196,7 @@ export const tryNativeToHexString = (
  */
 export const nativeToHexString = (
   address: string | undefined,
-  chain: ChainId
+  chain: ChainId | ChainName
 ): string | null => {
   if (!address) {
     return null;
@@ -202,8 +213,9 @@ export const nativeToHexString = (
  */
 export function tryNativeToUint8Array(
   address: string,
-  chainId: ChainId
+  chain: ChainId | ChainName
 ): Uint8Array {
+  const chainId = coalesceChainId(chain);
   return hexToUint8Array(tryNativeToHexString(address, chainId));
 }