+++ sdk/js/src/utils/consts.ts
@@ -39,6 +39,200 @@ export type EVMChainName =
   | "celo"
   | "ropsten";
 
+export type Contracts = {
+  core: string | undefined;
+  token_bridge: string | undefined;
+  nft_bridge: string | undefined;
+};
+
+export type ChainContracts = {
+  [chain in ChainName]: Contracts;
+};
+
+const MAINNET = {
+  unset: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  solana: {
+    core: "worm2ZoG2kUd4vFXhvjh93UUH596ayRfgQ2MgjNMTth",
+    token_bridge: "wormDTUJ6AWPNvk59vGQbDvGJmqbDTdgWgAqcLBCgUb",
+    nft_bridge: "WnFt12ZrnzZrFZkt2xsNsaNWoQribnuQ5B5FrDbwDhD",
+  },
+  ethereum: {
+    core: "0x98f3c9e6E3fAce36bAAd05FE09d375Ef1464288B",
+    token_bridge: "0x3ee18B2214AFF97000D974cf647E7C347E8fa585",
+    nft_bridge: "0x6FFd7EdE62328b3Af38FCD61461Bbfc52F5651fE",
+  },
+  terra: {
+    core: "terra1dq03ugtd40zu9hcgdzrsq6z2z4hwhc9tqk2uy5",
+    token_bridge: "terra10nmmwe8r3g99a9newtqa7a75xfgs2e8z87r2sf",
+    nft_bridge: undefined,
+  },
+  bsc: {
+    core: "0x98f3c9e6E3fAce36bAAd05FE09d375Ef1464288B",
+    token_bridge: "0xB6F6D86a8f9879A9c87f643768d9efc38c1Da6E7",
+    nft_bridge: "0x5a58505a96D1dbf8dF91cB21B54419FC36e93fdE",
+  },
+  polygon: {
+    core: "0x7A4B5a56256163F07b2C80A7cA55aBE66c4ec4d7",
+    token_bridge: "0x5a58505a96D1dbf8dF91cB21B54419FC36e93fdE",
+    nft_bridge: "0x90BBd86a6Fe93D3bc3ed6335935447E75fAb7fCf",
+  },
+  avalanche: {
+    core: "0x54a8e5f9c4CbA08F9943965859F6c34eAF03E26c",
+    token_bridge: "0x0e082F06FF657D94310cB8cE8B0D9a04541d8052",
+    nft_bridge: "0xf7B6737Ca9c4e08aE573F75A97B73D7a813f5De5",
+  },
+  oasis: {
+    core: "0xfE8cD454b4A1CA468B57D79c0cc77Ef5B6f64585",
+    token_bridge: "0x5848C791e09901b40A9Ef749f2a6735b418d7564",
+    nft_bridge: "0x04952D522Ff217f40B5Ef3cbF659EcA7b952a6c1",
+  },
+  algorand: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  fantom: {
+    core: "0x126783A6Cb203a3E35344528B26ca3a0489a1485",
+    token_bridge: "0x7C9Fc5741288cDFdD83CeB07f3ea7e22618D79D2",
+    nft_bridge: "0xA9c7119aBDa80d4a4E0C06C8F4d8cF5893234535",
+  },
+  aurora: {
+    core: "0xa321448d90d4e5b0A732867c18eA198e75CAC48E",
+    token_bridge: "0x51b5123a7b0F9b2bA265f9c4C8de7D78D52f510F",
+    nft_bridge: "0x6dcC0484472523ed9Cdc017F711Bcbf909789284",
+  },
+  karura: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  acala: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  klaytn: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  ropsten: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+};
+
+const TESTNET = {
+  unset: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  solana: {
+    core: "3u8hJUVTA4jH1wYAyUur7FFZVQ8H635K3tSHHF4ssjQ5",
+    token_bridge: "DZnkkTmCiFWfYTfT41X3Rd1kDgozqzxWaHqsw6W4x2oe",
+    nft_bridge: "2rHhojZ7hpu1zA91nvZmT8TqWWvMcKmmNBCr2mKTtMq4",
+  },
+  terra: {
+    core: "terra1pd65m0q9tl3v8znnz5f5ltsfegyzah7g42cx5v",
+    token_bridge: "terra1pseddrv0yfsn76u4zxrjmtf45kdlmalswdv39a",
+    nft_bridge: undefined,
+  },
+  ethereum: {
+    core: "0x706abc4E45D419950511e474C7B9Ed348A4a716c",
+    token_bridge: "0xF890982f9310df57d00f659cf4fd87e65adEd8d7",
+    nft_bridge: "0xD8E4C2DbDd2e2bd8F1336EA691dBFF6952B1a6eB",
+  },
+  bsc: {
+    core: "0x68605AD7b15c732a30b1BbC62BE8F2A509D74b4D",
+    token_bridge: "0x9dcF9D205C9De35334D646BeE44b2D2859712A09",
+    nft_bridge: "0xcD16E5613EF35599dc82B24Cb45B5A93D779f1EE",
+  },
+  polygon: {
+    core: "0x0CBE91CF822c73C2315FB05100C2F714765d5c20",
+    token_bridge: "0x377D55a7928c046E18eEbb61977e714d2a76472a",
+    nft_bridge: "0x51a02d0dcb5e52F5b92bdAA38FA013C91c7309A9",
+  },
+  avalanche: {
+    core: "0x7bbcE28e64B3F8b84d876Ab298393c38ad7aac4C",
+    token_bridge: "0x61E44E506Ca5659E6c0bba9b678586fA2d729756",
+    nft_bridge: "0xD601BAf2EEE3C028344471684F6b27E789D9075D",
+  },
+  oasis: {
+    core: "0xc1C338397ffA53a2Eb12A7038b4eeb34791F8aCb",
+    token_bridge: "0x88d8004A9BdbfD9D28090A02010C19897a29605c",
+    nft_bridge: "0xC5c25B41AB0b797571620F5204Afa116A44c0ebA",
+  },
+  algorand: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+  fantom: {
+    core: "0x1BB3B4119b7BA9dfad76B0545fb3F531383c3bB7",
+    token_bridge: "0x599CEa2204B4FaECd584Ab1F2b6aCA137a0afbE8",
+    nft_bridge: "0x63eD9318628D26BdCB15df58B53BB27231D1B227",
+  },
+  aurora: {
+    core: "0xBd07292de7b505a4E803CEe286184f7Acf908F5e",
+    token_bridge: "0xD05eD3ad637b890D68a854d607eEAF11aF456fba",
+    nft_bridge: "0x8F399607E9BA2405D87F5f3e1B78D950b44b2e24",
+  },
+  karura: {
+    core: "0xE4eacc10990ba3308DdCC72d985f2a27D20c7d03",
+    token_bridge: "0xd11De1f930eA1F7Dd0290Fe3a2e35b9C91AEFb37",
+    nft_bridge: "0x0A693c2D594292B6Eb89Cb50EFe4B0b63Dd2760D",
+  },
+  acala: {
+    core: "0x4377B49d559c0a9466477195C6AdC3D433e265c0",
+    token_bridge: "0xebA00cbe08992EdD08ed7793E07ad6063c807004",
+    nft_bridge: "0x96f1335e0AcAB3cfd9899B30b2374e25a2148a6E",
+  },
+  klaytn: {
+    core: "0x1830CC6eE66c84D2F177B94D544967c774E624cA",
+    token_bridge: "0xC7A13BE098720840dEa132D860fDfa030884b09A",
+    nft_bridge: "0x94c994fC51c13101062958b567e743f1a04432dE",
+  },
+  ropsten: {
+    core: undefined,
+    token_bridge: undefined,
+    nft_bridge: undefined,
+  },
+};
+
+/**
+ *
+ * If you get a type error here, it means that a chain you just added does not
+ * have an entry in TESTNET.
+ * This is implemented as an ad-hoc type assertion instead of a type annotation
+ * on TESTNET so that e.g.
+ *
+ * ```typescript
+ * TESTNET['solana'].core
+ * ```
+ * has type 'string' instead of 'string | undefined'.
+ *
+ * (Do not delete this declaration!)
+ */
+const isTestnetContracts: ChainContracts = TESTNET;
+
+/**
+ *
+ * See [[isTestnetContracts]]
+ */
+const isMainnetContracts: ChainContracts = MAINNET;
+
+/**
+ *
+ * Contracts addresses on testnet and mainnet
+ */
+export const CONTRACTS = { MAINNET, TESTNET };
+
 // We don't specify the types of the below consts to be [[ChainId]]. This way,
 // the inferred type will be a singleton (or literal) type, which is more precise and allows
 // typescript to perform context-sensitive narrowing when checking against them.
@@ -239,4 +433,12 @@ function isEVM(_: EVMChainId | EVMChainName): true {
  */
 function notEVM<T>(_: T extends EVMChainId | EVMChainName ? never : T): false {
   return false;
+}
+
+// This just serves as a type assertion to ensure that [[EVMChainName]] is a
+// subset of [[ChainName]], since typescript provides no built-in way to express
+// this.
+function evm_chain_subset(e: EVMChainName): ChainName {
+  // will fail to compile if 'e' can't be typed as a [[ChainName]]
+  return e;
 }